src/admin_cli/requirements-admin.txt
typer[all]==0.12.5
rich==13.9.4
psycopg[binary]>=3.2.10,<3.3
python-dotenv==1.0.1
